🚚 rename: issue #4 contentPost 구조 변경
