+++ src/main/java/com/example/streaming/model/ContentPost.java
@@ -1,6 +1,5 @@
-package com.example.streaming.domain;
+package com.example.streaming.model;
 
-import com.example.streaming.controller.request.ContentPostCreate;
 import lombok.Builder;
 import lombok.Getter;
