+++ src/main/java/com/example/streaming/model/ContentPostCreate.java
@@ -1,4 +1,4 @@
-package com.example.streaming.controller.request;
+package com.example.streaming.model;
 
 import lombok.Builder;
 import lombok.Getter;
