+++ src/main/java/com/example/streaming/repository/ContentPostJpaRepository.java
@@ -1,5 +1,6 @@
-package com.example.streaming.infrastructure;
+package com.example.streaming.repository;
 
+import com.example.streaming.common.entity.ContentPostEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ContentPostJpaRepository extends JpaRepository<ContentPostEntity, Long> {
