+++ src/main/java/com/example/streaming/service/ContentPostRepositoryImpl.java
@@ -1,7 +1,9 @@
-package com.example.streaming.infrastructure;
+package com.example.streaming.service;
 
-import com.example.streaming.domain.ContentPost;
-import com.example.streaming.service.port.ContentPostRepository;
+import com.example.streaming.common.entity.ContentPostEntity;
+import com.example.streaming.model.ContentPost;
+import com.example.streaming.repository.ContentPostJpaRepository;
+import com.example.streaming.repository.ContentPostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
